+++ android/src/main/java/com/instabug/reactlibrary/RNInstabugReactnativeModule.java
@@ -5,10 +5,11 @@ import static com.instabug.reactlibrary.utils.InstabugUtil.getMethod;
 import android.app.Application;
 import android.graphics.Bitmap;
 import android.net.Uri;
+import android.util.Log;
 import android.view.View;
 
-import androidx.annotation.UiThread;
 import androidx.annotation.NonNull;
+import androidx.annotation.UiThread;
 
 import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.Callback;
@@ -16,6 +17,7 @@ import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReactMethod;
 import com.facebook.react.bridge.ReadableArray;
+import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.bridge.WritableArray;
 import com.facebook.react.bridge.WritableMap;
 import com.facebook.react.bridge.WritableNativeArray;
@@ -29,6 +31,7 @@ import com.instabug.library.IssueType;
 import com.instabug.library.LogLevel;
 import com.instabug.library.ReproConfigurations;
 import com.instabug.library.core.InstabugCore;
+import com.instabug.library.featuresflags.model.IBGFeatureFlag;
 import com.instabug.library.internal.module.InstabugLocale;
 import com.instabug.library.invocation.InstabugInvocationEvent;
 import com.instabug.library.logging.InstabugLog;
@@ -36,6 +39,7 @@ import com.instabug.library.model.NetworkLog;
 import com.instabug.library.model.Report;
 import com.instabug.library.networkDiagnostics.model.NetworkDiagnosticsCallback;
 import com.instabug.library.ui.onboarding.WelcomeMessage;
+import com.instabug.library.util.InstabugSDKLogger;
 import com.instabug.reactlibrary.utils.ArrayUtil;
 import com.instabug.reactlibrary.utils.EventEmitterModule;
 import com.instabug.reactlibrary.utils.MainThreadHandler;
@@ -52,6 +56,7 @@ import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -64,7 +69,7 @@ import javax.annotation.Nullable;
  */
 public class RNInstabugReactnativeModule extends EventEmitterModule {
 
-    private static final String TAG = RNInstabugReactnativeModule.class.getSimpleName();
+    private static final String TAG = "IBG-RN-Core";
 
     private InstabugCustomTextPlaceHolder placeHolders;
     private static Report currentReport;
@@ -102,6 +107,7 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
 
     /**
      * Enables or disables Instabug functionality.
+     *
      * @param isEnabled A boolean to enable/disable Instabug.
      */
     @ReactMethod
@@ -110,7 +116,7 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
             @Override
             public void run() {
                 try {
-                    if(isEnabled)
+                    if (isEnabled)
                         Instabug.enable();
                     else
                         Instabug.disable();
@@ -123,10 +129,11 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
 
     /**
      * Initializes the SDK.
-     * @param token The token that identifies the app. You can find it on your dashboard.
+     *
+     * @param token                 The token that identifies the app. You can find it on your dashboard.
      * @param invocationEventValues The events that invoke the SDK's UI.
-     * @param logLevel The level of detail in logs that you want to print.
-     * @param codePushVersion The Code Push version to be used for all reports.
+     * @param logLevel              The level of detail in logs that you want to print.
+     * @param codePushVersion       The Code Push version to be used for all reports.
      */
     @ReactMethod
     public void init(
@@ -152,8 +159,8 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
                         .setInvocationEvents(invocationEvents)
                         .setLogLevel(parsedLogLevel);
 
-                if(codePushVersion != null) {
-                    if(Instabug.isBuilt()) {
+                if (codePushVersion != null) {
+                    if (Instabug.isBuilt()) {
                         Instabug.setCodePushVersion(codePushVersion);
                     } else {
                         builder.setCodePushVersion(codePushVersion);
@@ -165,7 +172,7 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
     }
 
     @ReactMethod
-    public void setCodePushVersion(@Nullable String version) {
+    public void setCodePushVersion(@Nullable final String version) {
         MainThreadHandler.runOnMainThread(new Runnable() {
             @Override
             public void run() {
@@ -319,7 +326,7 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
      *
      * @param userEmail User's default email
      * @param userName  Username.
-     * @param userId User's ID
+     * @param userId    User's ID
      */
     @ReactMethod
     public void identifyUser(
@@ -739,15 +746,15 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
 
     private WritableMap convertFromHashMapToWriteableMap(HashMap hashMap) {
         WritableMap writableMap = new WritableNativeMap();
-        for(int i = 0; i < hashMap.size(); i++) {
+        for (int i = 0; i < hashMap.size(); i++) {
             Object key = hashMap.keySet().toArray()[i];
             Object value = hashMap.get(key);
-            writableMap.putString((String) key,(String) value);
+            writableMap.putString((String) key, (String) value);
         }
         return writableMap;
     }
 
-    private static JSONObject objectToJSONObject(Object object){
+    private static JSONObject objectToJSONObject(Object object) {
         Object json = null;
         JSONObject jsonObject = null;
         try {
@@ -764,13 +771,12 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
     private WritableArray convertArrayListToWritableArray(List arrayList) {
         WritableArray writableArray = new WritableNativeArray();
 
-        for(int i = 0; i < arrayList.size(); i++) {
+        for (int i = 0; i < arrayList.size(); i++) {
             Object object = arrayList.get(i);
 
-            if(object instanceof String) {
+            if (object instanceof String) {
                 writableArray.pushString((String) object);
-            }
-            else {
+            } else {
                 JSONObject jsonObject = objectToJSONObject(object);
                 writableArray.pushMap((WritableMap) jsonObject);
             }
@@ -800,7 +806,7 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
 
     @ReactMethod
     public void setReproStepsConfig(final String bugMode, final String sessionReplayMode,
-                                    final String anr, final String appHang, final String fatal, final String nonFatal, final String forceRestart,String oom
+                                    final String anr, final String appHang, final String fatal, final String nonFatal, final String forceRestart, String oom
 
     ) {
         MainThreadHandler.runOnMainThread(new Runnable() {
@@ -825,7 +831,6 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
                             .setIssueMode(IssueType.SessionReplay, resolvedSessionReplayMode);
 
 
-
                     Instabug.setReproConfigurations(config.build());
                 } catch (Exception e) {
                     e.printStackTrace();
@@ -838,7 +843,7 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
      * Shows the welcome message in a specific mode.
      *
      * @param welcomeMessageMode An enum to set the welcome message mode to
-      *                          live, or beta.
+     *                           live, or beta.
      */
     @ReactMethod
     public void showWelcomeMessageWithMode(final String welcomeMessageMode) {
@@ -860,7 +865,7 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
      * Sets the welcome message mode to live, beta or disabled.
      *
      * @param welcomeMessageMode An enum to set the welcome message mode to
-      *                          live, beta or disabled.
+     *                           live, beta or disabled.
      */
     @ReactMethod
     public void setWelcomeMessageMode(final String welcomeMessageMode) {
@@ -911,27 +916,38 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
         });
     }
 
-    /**
-     * Extracts HTTP connection properties. Request method, Headers, Date, Url and Response code
-     *
-     * @param jsonObject the JSON object containing all HTTP connection properties
-     * @throws JSONException
-     */
     @ReactMethod
-    public void networkLog(String jsonObject) throws JSONException {
-        NetworkLog networkLog = new NetworkLog();
-        String date = System.currentTimeMillis()+"";
-        networkLog.setDate(date);
-        JSONObject newJSONObject = new JSONObject(jsonObject);
-        networkLog.setUrl(newJSONObject.getString("url"));
-        networkLog.setRequest(newJSONObject.getString("requestBody"));
-        networkLog.setResponse(newJSONObject.getString("responseBody"));
-        networkLog.setMethod(newJSONObject.getString("method"));
-        networkLog.setResponseCode(newJSONObject.getInt("responseCode"));
-        networkLog.setRequestHeaders(newJSONObject.getString("requestHeaders"));
-        networkLog.setResponseHeaders(newJSONObject.getString("responseHeaders"));
-        networkLog.setTotalDuration(newJSONObject.getLong("duration"));
-        networkLog.insert();
+    public void networkLogAndroid(final String url,
+                                  final String requestBody,
+                                  final String responseBody,
+                                  final String method,
+                                  final double responseCode,
+                                  final String requestHeaders,
+                                  final String responseHeaders,
+                                  final double duration) {
+        try {
+            final String date = String.valueOf(System.currentTimeMillis());
+
+            NetworkLog networkLog = new NetworkLog();
+            networkLog.setDate(date);
+            networkLog.setUrl(url);
+            networkLog.setMethod(method);
+            networkLog.setResponseCode((int) responseCode);
+            networkLog.setTotalDuration((long) duration);
+
+            try {
+                networkLog.setRequest(requestBody);
+                networkLog.setResponse(responseBody);
+                networkLog.setRequestHeaders(requestHeaders);
+                networkLog.setResponseHeaders(responseHeaders);
+            } catch (OutOfMemoryError | Exception exception) {
+                Log.d(TAG, "Error: " + exception.getMessage() + "while trying to set network log contents (request body, response body, request headers, and response headers).");
+            }
+
+            networkLog.insert();
+        } catch (OutOfMemoryError | Exception exception) {
+            Log.d(TAG, "Error: " + exception.getMessage() + "while trying to insert a network log");
+        }
     }
 
     @UiThread
@@ -984,7 +1000,6 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
      * Reports that the screen name been changed (Current View).
      *
      * @param screenName string containing the screen name
-     *
      */
     @ReactMethod
     public void reportCurrentViewChange(final String screenName) {
@@ -1007,7 +1022,6 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
      * Reports that the screen has been changed (Repro Steps) the screen sent to this method will be the 'current view' on the dashboard
      *
      * @param screenName string containing the screen name
-     *
      */
     @ReactMethod
     public void reportScreenChange(final String screenName) {
@@ -1017,7 +1031,7 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
                 try {
                     Method method = getMethod(Class.forName("com.instabug.library.Instabug"), "reportScreenChange", Bitmap.class, String.class);
                     if (method != null) {
-                        method.invoke(null , null, screenName);
+                        method.invoke(null, null, screenName);
                     }
                 } catch (Exception e) {
                     e.printStackTrace();
@@ -1026,6 +1040,9 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
         });
     }
 
+    /**
+     * @deprecated see {@link #addFeatureFlags(ReadableMap)}
+     */
     @ReactMethod
     public void addExperiments(final ReadableArray experiments) {
         MainThreadHandler.runOnMainThread(new Runnable() {
@@ -1042,6 +1059,9 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
         });
     }
 
+    /**
+     * @deprecated see {@link #removeFeatureFlags(ReadableArray)}
+     */
     @ReactMethod
     public void removeExperiments(final ReadableArray experiments) {
         MainThreadHandler.runOnMainThread(new Runnable() {
@@ -1058,6 +1078,9 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
         });
     }
 
+    /**
+     * @deprecated see {@link #removeAllFeatureFlags()}
+     */
     @ReactMethod
     public void clearAllExperiments() {
         MainThreadHandler.runOnMainThread(new Runnable() {
@@ -1065,6 +1088,59 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
             public void run() {
                 try {
                     Instabug.clearAllExperiments();
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+            }
+        });
+    }
+
+    @ReactMethod
+    public void addFeatureFlags(final ReadableMap featureFlagsMap) {
+        MainThreadHandler.runOnMainThread(new Runnable() {
+            @Override
+            public void run() {
+                try {
+                    Iterator<Map.Entry<String, Object>> iterator = featureFlagsMap.getEntryIterator();
+                    ArrayList<IBGFeatureFlag> featureFlags = new ArrayList<>();
+                    while (iterator.hasNext()) {
+                        Map.Entry<String, Object> item = iterator.next();
+                        String variant = (String) item.getValue();
+                        String name = item.getKey();
+                        featureFlags.add(new IBGFeatureFlag(name, variant.isEmpty() ? null : variant));
+                    }
+                    if (!featureFlags.isEmpty()) {
+                        Instabug.addFeatureFlags(featureFlags);
+                    }
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+            }
+        });
+    }
+
+    @ReactMethod
+    public void removeFeatureFlags(final ReadableArray featureFlags) {
+        MainThreadHandler.runOnMainThread(new Runnable() {
+            @Override
+            public void run() {
+                try {
+                    ArrayList<String> stringArray = ArrayUtil.parseReadableArrayOfStrings(featureFlags);
+                    Instabug.removeFeatureFlag(stringArray);
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+            }
+        });
+    }
+
+    @ReactMethod
+    public void removeAllFeatureFlags() {
+        MainThreadHandler.runOnMainThread(new Runnable() {
+            @Override
+            public void run() {
+                try {
+                    Instabug.removeAllFeatureFlags();
                 } catch (Exception e) {
                     e.printStackTrace();
                 }
@@ -1123,7 +1199,7 @@ public class RNInstabugReactnativeModule extends EventEmitterModule {
      * Map between the exported JS constant and the arg key in {@link ArgsRegistry}.
      * The constant name and the arg key should match to be able to resolve the
      * constant with its actual value from the {@link ArgsRegistry} maps.
-     *
+     * <p>
      * This is a workaround, because RN cannot resolve enums in the constants map.
      */
     @Override