+++ ios/RNInstabug/ArgsRegistry.h
@@ -1,5 +1,6 @@
 #import <Foundation/Foundation.h>
 #import <Instabug/IBGTypes.h>
+#import <Instabug/Instabug.h>
 
 typedef NSDictionary<NSString*, NSNumber*> ArgsDictionary;
 
@@ -20,6 +21,9 @@ typedef NSDictionary<NSString*, NSNumber*> ArgsDictionary;
 + (ArgsDictionary *) extendedBugReportStates;
 + (ArgsDictionary *) reproStates;
 + (ArgsDictionary *) locales;
++ (ArgsDictionary *)nonFatalExceptionLevel;
++ (ArgsDictionary *) launchType;
+
 + (NSDictionary<NSString *, NSString *> *) placeholders;
 
 @end