+++ ios/RNInstabug/InstabugCrashReportingBridge.m
@@ -1,6 +1,6 @@
 #import "InstabugCrashReportingBridge.h"
-#import <Instabug/IBGCrashReporting.h>
-#import <Instabug/Instabug.h>
+#import "Util/IBGCrashReporting+CP.h"
+
 
 @implementation InstabugCrashReportingBridge
 
@@ -29,30 +29,25 @@ RCT_EXPORT_METHOD(setEnabled: (BOOL) isEnabled) {
 RCT_EXPORT_METHOD(sendJSCrash:(NSDictionary *)stackTrace
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject) {
-    dispatch_queue_t queue = dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_HIGH, 0ul);
+    dispatch_queue_t queue = dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0ul);
     dispatch_async(queue, ^{
-        SEL reportCrashWithStackTraceSEL = NSSelectorFromString(@"reportCrashWithStackTrace:handled:");
-        if ([[Instabug class] respondsToSelector:reportCrashWithStackTraceSEL]) {
-            [[Instabug class] performSelector:reportCrashWithStackTraceSEL withObject:stackTrace withObject:@(NO)];
-        }
+        [IBGCrashReporting cp_reportFatalCrashWithStackTrace:stackTrace];
         resolve([NSNull null]);
     });
 }
 
 RCT_EXPORT_METHOD(sendHandledJSCrash:(NSDictionary *)stackTrace
+                  userAttributes:(nullable NSDictionary *)userAttributes fingerprint:(nullable NSString *)fingerprint nonFatalExceptionLevel:(IBGNonFatalLevel)nonFatalExceptionLevel
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject) {
     dispatch_queue_t queue = dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_HIGH, 0ul);
     dispatch_async(queue, ^{
-        SEL reportCrashWithStackTraceSEL = NSSelectorFromString(@"reportCrashWithStackTrace:handled:");
-        if ([[Instabug class] respondsToSelector:reportCrashWithStackTraceSEL]) {
-            [[Instabug class] performSelector:reportCrashWithStackTraceSEL withObject:stackTrace withObject:@(YES)];
-        }
+        [IBGCrashReporting cp_reportNonFatalCrashWithStackTrace:stackTrace level:nonFatalExceptionLevel groupingString:fingerprint userAttributes:userAttributes];
+
         resolve([NSNull null]);
     });
+
 }
-
-
 @synthesize description;
 
 @synthesize hash;